Orientation printing and basic instruction
For now, it can print its orientation (tilt up, tilt down, turing left, turing right or smooth flight) to the uart.

Also, if the board dropped to the ground, it will break the loop which means plane crashed.

I also enable the double-tap fuction on  the sensor to simulate the plane was hit by double tapping the sensor. If the board was hit 5 times, it will also break the loop which means the plane was shot down. And push button simulates the missiles fired.

But haven' t figured out how to measure the rising and falling status effectivelly.